requirements.txt
# telebot
pyTelegramBotAPI
requests
Pillow
